requirements.txt
azureml-sdk==1.54.0
numpy==1.23.5
pandas==2.1.2
scikit-learn==1.3.2